requirements.txt
# AI Year-In-Review Analysis Script Dependencies

# 中文分词（必需，用于中文关键词提取）
jieba>=0.42.1

# 英文自然语言处理（可选，用于英文分词和停用词）
nltk>=3.8

# 语言检测（可选，用于检测中英文）
langdetect>=1.0.9

# FastAPI Web 应用依赖
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.0
pydantic>=2.5.0

# DuckDuckGo 搜索功能
duckduckgo-search>=4.0.0

# URL 访问和解析
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

